:sparkles: (Pokemon - Create/Post) Implementing the possibility to create a new pokemon.
It is now possible to create a new Pokemon. As my database has already defined "is_legendary" as a string, it is only possible to input it that way.
Therefore, a LegendaryException was created to verify it. A middleware was also created to create new data to save the Pokedex number. This middleware
will only check once after initializing the application and making a request. With the creation of the middleware and the registration in a new
collection, it avoids the duplication of the Pokedex number in case any Pokemon is deleted in the future or to set the same Pokedex number of  pokemon
that was deleted earlier. The counter starts with number 801 (total number of pokemons until now).